requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.3
google-generativeai==0.8.0
google-genai>=1.0.0
playwright==1.41.0
pillow==10.2.0
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.0
